Merge remote-tracking branch 'origin/techArtsPerformance' into HEAD

Conflicts:
	JCocoSketch/src/jcocosketch/Line_Mod.java

+++ JCocoSketch/src/jcocosketch/QuadTree.java
@@ -47,7 +47,7 @@ public class QuadTree {
         if (x < root.getX() || y < root.getY() || x > root.getX() + root.getW() || y > root.getY() + root.getH()) {
             throw new QuadTreeException("Out of bounds : (" + x + ", " + y + ")");
         }
-        if (this.insert(root, new Point(x, y, value))) {
+        if (this.insert(root, new Point((float)x, (float)y, value))) {
             this.count_++;
         }
     }
